+++ tests/unit/project_loader/test_config.py
@@ -19,7 +19,6 @@ from textwrap import dedent
 from testtools.matchers import Contains, Equals, Is
 
 from . import LoadPartBaseTest, ProjectLoaderBaseTest
-from snapcraft.internal import project_loader
 
 from snapcraft.internal.project_loader import errors
 import snapcraft.internal.project_loader._config as _config
@@ -262,209 +261,3 @@ class FilesetsTest(unit.TestCase):
         self.assertThat(str(raised), Contains(
             '\'$3\' referred to in the \'stage\' fileset but it is not '
             'in filesets'))
-
-
-class VariableReplacementsTest(unit.TestCase):
-
-    def test_string_replacements(self):
-        replacements = (
-            (
-                'no replacement',
-                'snapcraft_stage/usr/bin',
-                'snapcraft_stage/usr/bin',
-            ),
-            (
-                'replaced start',
-                '$SNAPCRAFT_STAGE/usr/bin',
-                '{}/usr/bin'.format(self.stage_dir),
-            ),
-            (
-                'replaced between',
-                '--with-swig $SNAPCRAFT_STAGE/usr/swig',
-                '--with-swig {}/usr/swig'.format(self.stage_dir),
-            ),
-            (
-                'project replacement',
-                '$SNAPCRAFT_PROJECT_NAME-$SNAPCRAFT_PROJECT_VERSION',
-                'project_name-version',
-            ),
-        )
-
-        for test_name, subject, expected in replacements:
-            self.assertThat(project_loader.replace_attr(
-                subject, {
-                    '$SNAPCRAFT_PROJECT_NAME': 'project_name',
-                    '$SNAPCRAFT_PROJECT_VERSION': 'version',
-                    '$SNAPCRAFT_STAGE': self.stage_dir,
-                }), Equals(expected))
-
-    def test_lists_with_string_replacements(self):
-        replacements = (
-            (
-                'no replacement',
-                [
-                    'snapcraft_stage/usr/bin',
-                    '/usr/bin',
-                ],
-                [
-                    'snapcraft_stage/usr/bin',
-                    '/usr/bin',
-                ],
-            ),
-            (
-                'replaced start',
-                [
-                    '$SNAPCRAFT_STAGE/usr/bin',
-                    '/usr/bin',
-                ],
-                [
-                    '{}/usr/bin'.format(self.stage_dir),
-                    '/usr/bin',
-                ],
-            ),
-            (
-                'replaced between',
-                [
-                    '--without-python',
-                    '--with-swig $SNAPCRAFT_STAGE/usr/swig',
-                ],
-                [
-                    '--without-python',
-                    '--with-swig {}/usr/swig'.format(self.stage_dir),
-                ],
-            ),
-        )
-
-        for test_name, subject, expected in replacements:
-            self.assertThat(project_loader.replace_attr(
-                subject, {
-                    '$SNAPCRAFT_PROJECT_NAME': 'project_name',
-                    '$SNAPCRAFT_PROJECT_VERSION': 'version',
-                    '$SNAPCRAFT_STAGE': self.stage_dir,
-                }), Equals(expected))
-
-    def test_tuples_with_string_replacements(self):
-        replacements = (
-            (
-                'no replacement',
-                (
-                    'snapcraft_stage/usr/bin',
-                    '/usr/bin',
-                ),
-                [
-                    'snapcraft_stage/usr/bin',
-                    '/usr/bin',
-                ],
-            ),
-            (
-                'replaced start',
-                (
-                    '$SNAPCRAFT_STAGE/usr/bin',
-                    '/usr/bin',
-                ),
-                [
-                    '{}/usr/bin'.format(self.stage_dir),
-                    '/usr/bin',
-                ],
-            ),
-            (
-                'replaced between',
-                (
-                    '--without-python',
-                    '--with-swig $SNAPCRAFT_STAGE/usr/swig',
-                ),
-                [
-                    '--without-python',
-                    '--with-swig {}/usr/swig'.format(self.stage_dir),
-                ],
-            ),
-        )
-
-        for test_name, subject, expected in replacements:
-            self.assertThat(project_loader.replace_attr(
-                subject, {
-                    '$SNAPCRAFT_PROJECT_NAME': 'project_name',
-                    '$SNAPCRAFT_PROJECT_VERSION': 'version',
-                    '$SNAPCRAFT_STAGE': self.stage_dir,
-                }), Equals(expected))
-
-    def test_dict_with_string_replacements(self):
-        replacements = (
-            (
-                'no replacement',
-                {
-                    '1': 'snapcraft_stage/usr/bin',
-                    '2': '/usr/bin',
-                },
-                {
-                    '1': 'snapcraft_stage/usr/bin',
-                    '2': '/usr/bin',
-                },
-            ),
-            (
-                'replaced start',
-                {
-                    '1': '$SNAPCRAFT_STAGE/usr/bin',
-                    '2': '/usr/bin',
-                },
-                {
-                    '1': '{}/usr/bin'.format(self.stage_dir),
-                    '2': '/usr/bin',
-                },
-            ),
-            (
-                'replaced between',
-                {
-                    '1': '--without-python',
-                    '2': '--with-swig $SNAPCRAFT_STAGE/usr/swig',
-                },
-                {
-                    '1': '--without-python',
-                    '2': '--with-swig {}/usr/swig'.format(self.stage_dir),
-                },
-            ),
-        )
-
-        for test_name, subject, expected in replacements:
-            self.assertThat(project_loader.replace_attr(
-                subject, {
-                    '$SNAPCRAFT_PROJECT_NAME': 'project_name',
-                    '$SNAPCRAFT_PROJECT_VERSION': 'version',
-                    '$SNAPCRAFT_STAGE': self.stage_dir,
-                }), Equals(expected))
-
-    def test_string_replacement_with_complex_data(self):
-        subject = {
-            'filesets': {
-                'files': [
-                    'somefile',
-                    '$SNAPCRAFT_STAGE/file1',
-                    'SNAPCRAFT_STAGE/really',
-                ]
-            },
-            'configflags': [
-                '--with-python',
-                '--with-swig $SNAPCRAFT_STAGE/swig',
-            ],
-        }
-
-        expected = {
-            'filesets': {
-                'files': [
-                    'somefile',
-                    '{}/file1'.format(self.stage_dir),
-                    'SNAPCRAFT_STAGE/really',
-                ]
-            },
-            'configflags': [
-                '--with-python',
-                '--with-swig {}/swig'.format(self.stage_dir),
-            ],
-        }
-
-        self.assertThat(project_loader.replace_attr(
-            subject, {
-                '$SNAPCRAFT_PROJECT_NAME': 'project_name',
-                '$SNAPCRAFT_PROJECT_VERSION': 'version',
-                '$SNAPCRAFT_STAGE': self.stage_dir,
-            }), Equals(expected))
